🚚 rename use case to 'find-stream-links'
