+++ src/common/services/types.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, JSX } from 'preact'
-import { Metadata } from '../../release/utils/page-data'
+import { Metadata } from '../../find-stream-links/utils/page-data'
 
 export const SERVICE_IDS = [
   'applemusic',
